adding flag to build dkpg

+++ buildfw.sh
@@ -4,6 +4,7 @@ if [ $# -eq 0 ]
     echo "No arguments supplied. version number needed"
     exit
 fi
+
 Version=$1
 
 PROJECT_ROOT=`pwd`
@@ -36,32 +37,36 @@ echo "Done building genrate bdg.fw"
 
 mv $CAM_FW_ROOT/BirdDog_CAM_$Version.fw $PROJECT_ROOT
 
-#run build_sil2_webui_api_rapp.sh $Version sil2 webui api rapp
-#To create debian package
-#copy all $\BirdDog_CAM_FW\birddog-platform-files
+if [ $2 == "dpkgon" ];then
+    echo "build dpkg on"
+    #run build_sil2_webui_api_rapp.sh $Version sil2 webui api rapp
+    #To create debian package
+    #copy all $\BirdDog_CAM_FW\birddog-platform-files
+    
+    mkdir -p $DEB_SOURCE_PATH && cp -rf $BIRDDOG_PLATFORM_FILES_PATH/* $_
+    cp -r $PROJECT_ROOT/DEBIAN $DEB_SOURCE_PATH
+    echo "Preparing  debian file system."
+    echo "Package: BirdDog-Firmware
+    Version: $Version
+    Section: frimware
+    Priority: optional
+    Architecture: armhf
+    Maintainer: Anshul Yadav (anshuly@birddog.tv)
+    Description: This is Bird dog camera firmware.
+    " > $DEB_SOURCE_PATH/DEBIAN/control
+    
+    #create and copy all temp mcu bin and etc
+    mkdir -p $DEB_SOURCE_PATH/tmp && cp -f $DEB_ROOT_PATH/external/* $_
+    
+    cd $DEB_ROOT_PATH
+    chmod 0755 $DPKG_SCRIPT_PATH/preinst
+    chmod 0755 $DPKG_SCRIPT_PATH/postinst
+    chmod 0755 $DPKG_SCRIPT_PATH/control
+    
+    dpkg-deb --build $DEB_SOURCE_PATH
+    echo "Done building debian-package."
+   
+    rm -rf $DEB_SOURCE_PATH
+fi
 
-mkdir -p $DEB_SOURCE_PATH && cp -rf $BIRDDOG_PLATFORM_FILES_PATH/* $_
-cp -r $PROJECT_ROOT/DEBIAN $DEB_SOURCE_PATH
-echo "copied debian folder"
-
-echo "Package: BirdDog-Firmware
-Version: $Version
-Section: frimware
-Priority: optional
-Architecture: armhf
-Maintainer: Anshul Yadav (anshuly@birddog.tv)
-Description: This is Bird dog camera firmware.
-" > $DEB_SOURCE_PATH/DEBIAN/control
-
-#create and copy all temp mcu bin and etc
-mkdir -p $DEB_SOURCE_PATH/tmp && cp -f $DEB_ROOT_PATH/external/* $_
-
-cd $DEB_ROOT_PATH
-chmod 0755 $DPKG_SCRIPT_PATH/preinst
-chmod 0755 $DPKG_SCRIPT_PATH/postinst
-chmod 0755 $DPKG_SCRIPT_PATH/control
-
-dpkg-deb --build $DEB_SOURCE_PATH
-
-rm -rf $DEB_SOURCE_PATH
